refactor: extract queries build logic to split in shorter files
